skin cancer detector/requirements.txt
streamlit==1.31.1
tensorflow==2.15.0
numpy==1.26.0
pillow==10.2.0
opencv-python==4.9.0.80
h5py==3.10.0
protobuf==4.25.1
